refactor(SA2-74): Period/日付ドメインを shared/lib へ昇格＋Displayラベル統一
レビュー指摘対応:
- #2 (横断依存): sessions が statistics 配下から直接参照していた Period/日付
  ドメイン（resolveDateRange / dateInputToEpochSec / epochSecToDateInput /
  PERIODS / PERIOD_LABEL）を shared/lib/period-filter.ts へ昇格。statistics・
  sessions 双方が shared を参照する形にし、feature→feature 依存を解消。
  対象関数のユニットテストも shared 側 (__tests__/period-filter.test.ts) へ移設。
- #1 (ラベル不一致): Display チップ "BB · BI" とシート選択肢 "BB / BI" の表記
  揺れを、session-filters-helpers の SESSION_DISPLAY_LABEL 一本で両方を描く形に
  統一（"BB / BI"）。
