+++ apps/web/src/features/statistics/utils/stats-filters.ts
@@ -1,16 +1,7 @@
 import z from "zod";
+import { PERIODS, resolveDateRange } from "@/shared/lib/period-filter";
 
 // ── Filter value domains ─────────────────────────────────────────────────────
-
-export const STATS_PERIODS = [
-	"7d",
-	"30d",
-	"90d",
-	"ytd",
-	"all",
-	"custom",
-] as const;
-export type StatsPeriod = (typeof STATS_PERIODS)[number];
 
 // Normalization is a single mode. "off" shows currency amounts (a currency
 // must be selected); "normalized" shows big-blind (cash) and buy-in
@@ -23,12 +14,11 @@ export type StatsNormalization = (typeof STATS_NORMALIZATIONS)[number];
 export const STATS_TYPES = ["all", "cash_game", "tournament"] as const;
 export type StatsType = (typeof STATS_TYPES)[number];
 
-const DAY_SECONDS = 86_400;
-
 // ── URL search-param schema (route `validateSearch`) ─────────────────────────
 
 export const statsSearchSchema = z.object({
-	period: z.enum(STATS_PERIODS).default("all"),
+	// Preset / custom date window — shared with the sessions filter.
+	period: z.enum(PERIODS).default("all"),
 	// Custom-range bounds, Unix seconds. Coerced because raw URL values arrive
 	// as strings on a cold load / shared link.
 	from: z.coerce.number().int().optional(),
@@ -59,55 +49,6 @@ export interface StatsQueryInput {
 	normalized: boolean;
 	roomId?: string;
 	type?: "cash_game" | "tournament";
-}
-
-function startOfUtcDaySec(nowSec: number): number {
-	const d = new Date(nowSec * 1000);
-	return Math.floor(
-		Date.UTC(d.getUTCFullYear(), d.getUTCMonth(), d.getUTCDate()) / 1000
-	);
-}
-
-function startOfUtcYearSec(nowSec: number): number {
-	const d = new Date(nowSec * 1000);
-	return Math.floor(Date.UTC(d.getUTCFullYear(), 0, 1) / 1000);
-}
-
-/**
- * Translate the selected period into a concrete `{ dateFrom?, dateTo? }` window
- * in Unix seconds. Relative windows snap their bounds to UTC day boundaries so
- * the value (and therefore the query key) only changes once a day, not on every
- * render. The upper bound is the end of today so future-dated sessions are
- * excluded from "last N days" / YTD windows.
- */
-export function resolveDateRange(
-	filters: Pick<StatsFilters, "from" | "period" | "to">,
-	nowSec: number = Math.floor(Date.now() / 1000)
-): { dateFrom?: number; dateTo?: number } {
-	const dayStart = startOfUtcDaySec(nowSec);
-	const dayEnd = dayStart + DAY_SECONDS;
-	switch (filters.period) {
-		case "7d":
-			return { dateFrom: dayStart - 7 * DAY_SECONDS, dateTo: dayEnd };
-		case "30d":
-			return { dateFrom: dayStart - 30 * DAY_SECONDS, dateTo: dayEnd };
-		case "90d":
-			return { dateFrom: dayStart - 90 * DAY_SECONDS, dateTo: dayEnd };
-		case "ytd":
-			return { dateFrom: startOfUtcYearSec(nowSec), dateTo: dayEnd };
-		case "custom": {
-			const range: { dateFrom?: number; dateTo?: number } = {};
-			if (filters.from !== undefined) {
-				range.dateFrom = filters.from;
-			}
-			if (filters.to !== undefined) {
-				range.dateTo = filters.to;
-			}
-			return range;
-		}
-		default:
-			return {};
-	}
 }
 
 /** Map UI filter state to the tRPC `stats.*` query input. */
@@ -157,35 +98,4 @@ export function statsUnitFor(
 	currencyUnit: string | null | undefined
 ): string | null {
 	return norm === "off" ? (currencyUnit ?? null) : normalizedUnitForType(type);
-}
-
-// ── Custom-range date <input type="date"> conversion ─────────────────────────
-
-const DATE_INPUT_RE = /^\d{4}-\d{2}-\d{2}$/;
-
-/**
- * Convert a `yyyy-mm-dd` value from a date input into Unix seconds (UTC).
- * `endOfDay` snaps to 23:59:59 so an upper bound is inclusive of the whole day.
- * Returns undefined for an empty / malformed value so the bound is cleared.
- */
-export function dateInputToEpochSec(
-	value: string,
-	endOfDay = false
-): number | undefined {
-	if (!DATE_INPUT_RE.test(value)) {
-		return undefined;
-	}
-	const [y, m, d] = value.split("-").map(Number);
-	const ms = endOfDay
-		? Date.UTC(y, m - 1, d, 23, 59, 59)
-		: Date.UTC(y, m - 1, d, 0, 0, 0);
-	return Math.floor(ms / 1000);
-}
-
-/** Convert Unix seconds back to a `yyyy-mm-dd` value for a date input (UTC). */
-export function epochSecToDateInput(sec: number | undefined): string {
-	if (sec === undefined) {
-		return "";
-	}
-	return new Date(sec * 1000).toISOString().slice(0, 10);
 }
